CompareSignXAI/requirements.txt
matplotlib>=3.7.0
scipy>=1.10.0
version-parser>=1.0.1
tensorflow>=2.8.0
pytest>=7.0.0
pytest-cov>=4.0.0
torch>=2.0.0
zennit>=0.5.0
numpy>=1.20.0
pillow>=8.0.0